Add files via upload
I have taken on board the suggestions at the last meeting from Paul, Pedro and Rob. These are revised versions of the code flows to support the new FX functionality. They are simpler in themselves, and should also make it easier to implement in a common way between vNow and vNext